Added a Permissions Dialog
Added a permissions dialog in the Settings app.
This dialog shows all permissions from all installed apps,
and the apps which are using them.
When you click on a child entry (app) you will be redirected
to the InstalledAppDetails.

This is a first commit. Still some work to do with the layouts.
Please feel free to comment on this change. (This paragraph will
be removed in the final commit message)

Change-Id: If9b191770b8c5955ceb10ea1304153bb0bb0e35d