Battery Use 'Time on' label overlaps actual time value fixed
Both the 'Time on' label and actual time value is displayed on one
line. As the text increases due to language change(Espanol) and time
in days, hours, minutes, seconds, there is overlap of text in display.
Solution is to put 'Time on' label and actual time on seperate lines.

Change-Id: I06ff2852cb142fefdd7ade786664eda295ef4179