Fix preserving new APN with an unset "name" and "APN".
If an new APN have invalid values when the device goes
to sleep while editing, then the APN can't be saved when
the activity is resumed even though the values is corrected.
The behavior is identical when an other activity covers the
APN editor and onSaveStateInstance is called before "name"
and "APN" is set. Fixed so that the APN setting is
preserved despite of missing parameters "APN" and "name".

Change-Id: I5459efea65fbbdcde16995f35c9f0386ec8caae6