Update patch set 1
Patch Set 1: Verified

As far as I can tell the code merges and builds fine into Google's internal master source tree, and there are no build issues.

However, using a system property might turn out to be controversial, though it's not obvious to me in this specific case that there'd be a clear alternative.

Patch-set: 1