apps/Settings: Airplane mode.
Airplane mode depends on a cellular modem to return
a power off notification to the application to clear and
set the airplane mode database setting.

Systems without modems see that the intent return occurs before the
the setAirplane mode command is called thereby disabling the checkbox.

Only when the application looses focus and is re-entered is the checkbox
enabled and the command can be sent again.

This change will check to see if a cellular modem is available and if it
is send the intent to the modem.  If not the intent to the modem is not sent
but the intent is sent to the WiFi and Bluetooth.

Signed-off-by: Dan Murphy <dmurphy@ti.com>