Check battery level when doing master clear
If phone is resetting and battery is drained the phone might end up in
an unsupported state. If the battery level is low a dialog will alert
the user and master clear will not proceed. When user presses the
dialog button master clear will revert to its initial ui state.

Further changes include that the onBackPressed method is implemented
to show initial ui state if final state is active. This removes the
flashing of initial ui state when pressing back in final stage.