Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This doesn't merge cleanly into our internal tree due to some changes we made, but I can take care of the merge conflict.

This is a good change but I have a few suggestions to improve it further. We should add an OnDismissListener to the alert dialog so that it can call the unregisterDialog() method that you added when the user dismisses the dialog. Here's the code I added above showDisconnectDialog() to create the dismiss listener as a new field:

    private final DialogInterface.OnDismissListener mDismissListener =
            new DialogInterface.OnDismissListener() {
        public void onDismiss(DialogInterface dialog) {
            mLocalManager.unregisterDialog(dialog);
        }
    };

Then you can add:

    dialog.setOnDismissListener(mDismissListener);

after the line that creates the new dialog. Then the dismiss listener will be called to remove the dialog from the list in LocalBluetoothManager if the user dismisses it.

The problem with making this change is that the loop to dismiss the dialogs in the list in LocalBluetoothManager.setForegroundActivity() will cause the dismiss listener to be called, calling the unregisterDialog() method which will remove each dialog from the list as it is being iterated over, which is not allowed.

I suggest adding a single "mDisconnectDialog" field instead of a list to LocalBluetoothManager, because there should only ever be one disconnect dialog open at a time anyway. Then the register and unregister methods could look like:

    public void registerDialog(DialogInterface dialog) {
        mDisconnectDialog = dialog;
    }

    public void unregisterDialog(DialogInterface dialog) {
        if (mDisconnectDialog == dialog) {
            mDisconnectDialog = null;
        }
    }

and the cleanup added to setForegroundActivity() could look like:

        if (mDisconnectDialog != null) {
            // dismiss listener will set mDisconnectDialog to null
            mDisconnectDialog.dismiss();
        }

Also, both mErrorDialog and mDisconnectDialog can be changed from AlertDialog to DialogInterface since no AlertDialog-specific methods are called, and then the registerDialog() and unregisterDialog() methods can take a DialogInterface as a parameter (as in my versions above) and the dismiss listener doesn't need a cast to call the unregister method on the parameter that is passed to it.

Patch-set: 1