Fix NPE due to missing ProcessItem.mPackageInfo
The reason was that a item was removed from the SparseArray while
iterating it causing one ProcessItem to be skipped in the loop which
makes sure that mPackageInfo is not null. This happens when all
processes for one uid is stopped and a new process is created.
This problem was found by running monkey.

Change-Id: I5e9a76e8007819d5e6d9ba15af0c2362da193526