Limiting WPA password box to the standard.
In case of ascii the max length is 63 and in case of hex it is 64.

This fixes the below scenario:

If defining"Portable Wi-Fi hotspot" with WPA2 key containing 64
characters (or more) the hotspot will be deactivated and "Error" will
be shown. Then if you activate hotspot again and check settings, the
hotspot will be up and running and settings will say it is secured with
WPA2 eventhough it is activated as an "open" hotspot and others can
connect witout security check.

The setting will be updated first after stepping backwards in the menu
and then in again to see that the Hotspot is actually defined and active
as "open".

This could be really problematic for users that might be mislead to think
they have secured their Hotspot.

Change-Id: Idc0a1bf80f3e8b6db312e2cd9c3ba380c5fb15cf