Make sure dialog is not left hanging for invalid windows
Problem: LocalBluetoothManager is a singleton, which manages a set
of CachedBluetoothDevice, which can popup a dialog when disconnecting.
However, since this dialog is popped up in the context of a specific
BluetoothSettings instance - and we can get multiple instances f.ex.
when using "Home" to exit out of Settings and then restarting Settings
repeatedly - we must make sure that it's handled correctly both in
respect to which was its "parent activity" when it was created AND
in respect to switching orientation (we don't want to leak the dialog
window when rotating).

Change-Id: Iab1d5996462f2b91c25e78c7b3bdb850ac23c42d