Wrong behavior in the Settings application when the phone is rotated.
When the phone is rotated the focused activity is restarted. If the user sets a
new value in the Settings application and rotates the phone before saving, the
new value will be lost. Problems like this occur when changing the SIM PIN,
when changing the brightness level and when performing a factory data reset.

Change-Id: I1fbf78eb91e1db5d7cc4109b60226951b0391389