Making it possible to style Settings using overlay
Today it is not possible to style Settings in a good way using the
overlay mechanism. The reason for this is that Settings is missing
an application theme that can be overridden.

The solution is to add both a default application theme as well as
local styles for the themes used by specific Settings activities,
so that they can be overridden using overlay.

Change-Id: Ia084bbc09900c250bdfcb60a74b373ece18f5d37