Text labels when settings PIN/Password are not correct
Text labels were not consistent when setting screen lock
using PIN code or Password in settings. The user is getting
the wrong feedback (text9  in the UI when the incorrect
PIN/Password are inserted. For example if the user enters
too long password the UI will say "PIN can be no longer than
16 digits" and this is wrong.

Change-Id: Ia9295f7b36b29a3b4370e04461446a1428fc2ee2