WifiApDialog has wrong lifecycle
Best way to reproduce the problem:
1. Go to Settings -> Wireless -> Tethering.. -> Portable Wi-fi ..
 -> Configure ..
2. Click on Security so the picker dialog opens
3. Press home button
4. Go to Settings. (App crashes)

The problem is that the Spinner in WifiApDialog.java tries to dismiss
the dialog after the dialog has been removed by the window manager
because of leak detection.

The reason this happens is that the WifiApDialog that's created in
WifiApSettings.java is created in such a way that the life-cycle of
the Dialog is handled by the base class Activity. That causes the
dialog problem that the WifiApDialog is removed in onDestroy() of
WifiApSettings, which will send an event that will execute
onDetachFromWindow() on the Spinner instance in the future (this is
handled in ViewRoot.java). That message won't be processed until the
WindowManager has removed the view (the Spinner dialog) because of
leak detection, and once the onDetachedFromWindow() on Spinner
executes it dialog has already been removed.

The problem can be solved by handling the life-cycle of the
WifiApDialog completely in WifiApSettings analog to WifiSettings.

Change-Id: I6c5eea9d155589670f9b819efc28eb2e397e72ac