Wrong behavior in the Settings application when the phone is rotated.
This commit contains three fixes for rotation problems in the
Settings application:
1. IccLockSettings - changes in the code to handle the use case
"rotate the phone during the SIM PIN change process".
2. MasterClear - changes in the code to handle the use case
"rotate the phone during the Master Clear process".
3. Brightness - changes in the manifest to handle the use case
"rotate the phone when setting the brightnes level". This use case
was not possible to handle in the code the same way as the other
use cases since onDialogClosed(), for the Brightness Preference, is
called after the DisplaySettings Activity already is restarted due
to the rotation. Since the physical display brightness is always set
in onDialogClosed() a glitch can not be prevented unless the
manifest is modified.

Change-Id: I1fbf78eb91e1db5d7cc4109b60226951b0391389