Give the user a 'try again' button when PIN code fails at BT pair
When pairing with Bluetooth you are entering a PIN code but if
the PIN is wrong you are presented with an error dialog and when
pressing the ok button you will have to find your Bluetooth
device again in the UI list in order to retry. A big American
operator wants to have a retry button in the error dialog to
ease the process when you insert wrong PIN code, thus starting
the pairng dialog again when you press retry.

Change-Id: I6f9366cab917939e4cc5671fd5be6e7242dc02b2