Keyboard order in LanguageSettings
In Asia many languages have their own alphabets and these alphabets
are often best supported by a separate keyboard input method
implementation. Big operators require pre-installed support not only
for the mother tongue but also for minority languages as well as
a priority order of these keyboards.

This patch allows an array of flattened components to be put in an
overlay to define the order.

Change-Id: I5eb88e0b13ee6f7c679d0fbb0d782fd367201dbb