screen activation on pairing request
When the screen is off, the screen is not activated
on a pairing request notification from a bluetooth device.

This patch adds a full wakelock with a 15s timeout
when a pairing request that occurs when the screen
is off is detected. The effect is that the screen
goes bright for 15s when the request is received, leaving the
opportunity for the user to treat it properly.

Change-Id: I744cc37294db6fe631175ed1ef0929b2786096e7
Author: Ovidiu Beldie <ovidiux.beldie@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 3669 17106