Update patch set 1
Patch Set 1:

I like having the policy status show in the settings app, the way that other things do, such as System version, Model number, or Android version.

But it might be better to have a separate app, which can be included or excluded from a specific device, to actually change policy.  

I suspect that carriers would think that for consumer devices, the policy should always be enforcing.

Patch-set: 1