Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Unfortunately this probably isn't a change we can take, in this form. The API for the GPS provider and the corresponding uses-feature constant is defined with the assumption that the GPS hardware is built in to the device (i.e. is always present.) We can't take a change like this that makes the GPS hardware come and go, without violating the contract of the API.

We would need to introduce a notion of "always present" vs. "capable of" features -- and of course build out the framework plumbing -- before we could do something like this.

Patch-set: 1
Reviewer: Dan Morrill <1000962@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0