Add white and black list for disabling applications
This adds a whitelist so that it is possible to disable applications,
even though they are signed with the system certificate.
This adds a blacklist so that it is not possible to disable applications,
even though they reside in the data partition.

The lists are configurable through overlays in the product configuration.
There are separate lists for manufacturer and operator.

Change-Id: Ib257b88c57b1c870a708bc018c2e60b92a5635ce