Update patch set 1
Patch Set 1:

I think the current behavior is working as intended.

When the value is not in the list, we pick the closest higher value in the list. As you said this is an uncommon case anyway.

My point is there is no good solution here. I prefer the current solution that copes with rounding errors in the values and provides a 'best effort' answer rather than no information.

My main point is there should be no crash with this code unlike what is mentioned in the original bug report (note I think I changed this code during the ICS time frame, which may explain it).

If you agree, I would rather not merge this code.

Patch-set: 1