Update patch set 1
Patch Set 1: Verified

At a glance, the code style is fine. With the related changes, this merges and builds fine in Google's internal master branch. Untested.

Note to the author: you left a blank line after your Change-Id, so Gerrit ignored it and generated a new one.

Patch-set: 1
Label: Code-Review=0