Don't direct to market if no markets installed
The search intent cannot be resolved if there aren't any
market apps installed, so it will cause Settings crashed by
the ActivityNotFound exception.

So we need to check if the market intent requst can be resolved.
Then determine to notify user to direct the accessibility services
to market to get TalkBack or not.

Change-Id: I0f0d7b54eb5215e1a886315bb6816704384f6c59
Signed-off-by: Roger Chen <cxr514033970@gmail.com>

Conflicts:

	src/com/android/settings/AccessibilitySettings.java