Show default locale on top of list in LocalePicker
Currently the order of the locales are sorted alphabetically.
This change introduces adding top default locales. They can be either
customized to be visible on top or be selected from what language
the sim card reports. If neither of these is a valid locale default
locale will be the sw reported locale.

Furthermore the currently set locale is always added among the top
locales.

This makes it easier to find desired locales

Change-Id: Idab694c5efd73cad190fa3c989ca87faef9f08c5