Update patch set 5
Patch Set 5:

I get how it works, I really do :-) That's not the issue. So none of our code uses msgid, and if Google is doing it, do they have a tool for this to make it easier, why would I need to write one and are they required?

Patch-set: 5