Update patch set 1
Patch Set 1: Do not submit

I don't think this is the right approach.  If the device has actually gone to sleep and the device is locked then we shouldn't wake it up spuriously just because of the pairing request.

Typically the device will need to be in a discoverable mode in order to receive such a request.  That means the user is (hopefully) paying attention.

It might be ok to call "userActivity" to keep the screen on a little longer, but I don't think that's necessary either.

Patch-set: 1
Reviewer: Jeff Brown <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2