Settings: Disable SIM lock UIs in Airplane mode/SIM removal
If the airplane mode is activated using the power
button and if the current activity is one of the
SIM lock activities, then the SIM lock UI is not
disabled. Since the framework doesn't completely
support SIM access in airplane mode, there are few
issues seen. Same applies to if the SIM is removed
when the current activity is SIM lock activity.

This patch disables/greys out the SIM lock UIs when
the airplane mode is activated using the power button
or upon SIM removal.

Change-Id: Ib3b311da98d34392482669d253b396d066fe5e3a
Author: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 58530