Add Recife timezone (1/2)
Added Recife timezone due to in DST period a large part of northeast doesn't have time
correction and using Brasilia time it automatically changes clock in +1h affecting a lot
of applications that uses OS timezone setting as reference.

The northeast region has 9 states and represents 18% of the Brazilian territory.

Change-Id: Ib1a021a665ca09fcec8caf867efe4598c46045b0