Renaming Wifi direct device with <33 bytes
A user may not associate the Wifi direct device name
with the restrictions of SSID, and may thus use
any characters and arbitrary length when renaming.

Exceeding the limit (easy in asian languages),
displays a toast with "Failed to rename device."

Adds a Utf8 byte length filter, and avoids unnecessary
rename attempts.

Change-Id: I42339c0c170b9c550f9c061e4e40f22ec67a7e9f