Update patch set 4
Patch Set 4:

Bjorn Bringert: These sort of issues does indeed seem to occur  in several places. My guess is that these are automatically translated and then never reviewed or the person reviewing them doesn't know the context in which they are used.

Btw. I didn't know there were any Swedish people on the Android team. Am I correct in guessing that you don't work at the Stockholm office?

It would be great if someone with access to that internal system could verify if for me.

Patch-set: 4