Settings/Wifi: Add support for Ad-Hoc (IBSS) networks
If the WLAN driver can support IBSS mode, show Ad-Hoc (IBSS) networks in the
list of available Wi-Fi networks and allow to connect to or create new Ad-Hoc
networks.

This makes use of the new functions of WifiManager in frameworks/base. We have
to be careful to make the query only after we reached WIFI_STATE_ENABLED.
WifiSettingsForSetupWizardXL is left out because it does not handle these
states.

This requires the IBSS changes in frameworks/base.

Change-Id: Id4ebfd28cfd96df1f70f2f5bd35390d65dcb0c1f