Wifi: Enable IBSS options only if IBSS is supported
Show IBSS network options only if IBSS mode is supported by the Wifi driver.

This makes use of the new API isIbssSupported() of WifiManager. We have to be
careful to make the query only after we reached WIFI_STATE_ENABLED.

WifiSettingsForSetupWizardXL is left out because it does not handle these
states.

Change-Id: I463ebc58ab16494680c04b354c5c8f4d520aaceb