Modify display text for Mount/Unmount USB storage
Devices with support for USB storage devices in fstab configuration will
support USB drives as external storage devices and mount them properly
in Storage Settings, but the texts for these storages will be the same
as for SD card.
Add USB storage string in xml file for USB storage, and distinguish it from
the SD card.

Change-Id: I196bcd00b174ea41d3aeeef904009fa88e52fae0
Signed-off-by: juntingwang <Junting.Wang@windriver.com>
Signed-off-by: Robert Chiras <robert.chiras@intel.com>
Signed-off-by: Alexandru Timohi <alexandru.timohi@intel.com>
Signed-off-by: Irina Patru <irina.patru@intel.com>