Modify display text for Mount/Unmount USB storage
Devices with support for USB storage devices in fstab configuration will
support USB drives as external storage devices and mount them properly
in Storage Settings, but the texts for these storages will be the same
as for SD card.
Add USB storage string in xml file for USB storage, and distinguish it from
the SD card.

Change-Id: I0a17bb7f63bb675e271697752aca72fae09059dc
Signed-off-by: juntingwang <Junting.Wang@windriver.com>
Signed-off-by: Robert Chiras <robert.chiras@intel.com>