Issue #37178 Enable users to enter a string to match a certificate CN
In the current master branch, the new 802.11 api exposes the subject_match
setting from the wpa_supplicant configuration file, but does not expose
that functionality to the existing UI.  This patch enables that portion
of the API through the UI.  Specifically, the patch only enables matching
the CN portion of the certificate subject.  This behavior would put Android
in line with the behavior of other operating systems on the market.

Change-Id: Ib933a5f0687524319b504b39bc86837c80ab739a