Update patch set 1
Patch Set 1:

that wouldn't be the first time zone that isn't properly translated in Settings. i'd like to kill these translations by exposing some more icu4c API so i can have Settings do something like this instead:

 for each time zone Settings shows:
   if the time zone's country == the user's country:
     use the display name appropriate to the current time ("Eastern Standard Time", "Pacific Daylight Time")
   else:
     use the exemplar city ("Calcutta", "コルカタ")
     as an optional extra, i could add the translated country name too ("Calcutta (India)").

that seems to be roughly what the _intent_ of the Settings xml files was meant to be, even if that's not actually what we ended up with.

i'd been hoping to do this for the last two weeks, but other stuff keeps coming up. maybe tomorrow...

Patch-set: 1
Label: Code-Review=0