Update patch set 4
Patch Set 4:

Angela: It's quite possible that this has already been fixed internally in your translation console as Baligh said was needed. Since I never got any feedback here I just assumed it was still not fixed. After all this change was submitted back in September 2012.

Since TC is an internal system I can't verify it so I'll keep this change open until the next AOSP code drop and verify it then if that's ok. If you are sure this is no longer an issue I guess I can abandon it though.

Patch-set: 4