Update patch set 2
Patch Set 2:

Robert, if we keep original API CancelConnect intact and introduce a new API CancelDeviceConnect and call in WifiP2pSettings while cancel dialog listener is invoked, the original functionality of canceling the ongoing P2p connection would be replaced. 

Instead the new API CancelDeviceConnect may be invoked from wifiP2psettings when a group is formed (That said when a device is in group craeted state and and recieves a new connection request and is cancelled).

Could you please provide your comments

Patch-set: 2