Show available space on external media, immediately on mounting
The Storage settings page did not create the entries to display the Usage Meter, Total Space and Available Space on mounting external media. It created these entries only during the initialization of the page, which resulted in the user having to leave the page and reopen it to view the entries.

Bug: 8103899
Change-Id: I2b4777ad96091476810798957d418fc2e0cf3199