[BT] Add missing device upon incoming pairing request
When a device receives an incoming pairing request from a device
which has never been seen in an inquiry yet, this trigs an error
(error trace 'Got bonding state changed for <BD_ADDR>, but we have
no record of that device' + some callbacks are not called).

Solution: add this device to the map of 'cached devices' (which
for example encompasses all devices found by inquiries).

Change-Id: Idec69052bd615296baa5a1779fbff06bdc93618d
Signed-off-by: Enming XIE <enmingx.xie@intel.com>
Signed-off-by: Abdelmajid MLAYEH <abdelmajidx.mlayeh@intel.com>