Update patch set 1
Patch Set 1:

yamasani: are you happy with this? i want to move on to fixing the bidi bug.

(we're talking with sblitz about moving to Calendar's fancy time zone picker, but i'd like to get this in anyway just in case that doesn't make the next release. at the moment, several of our supported languages just get English strings here.)

Patch-set: 1