usb_tethering: Remove to disable menu
When MTP and PTP and USB Debugging mode is disabled, USB tethering checkbox
does not be activated although there is no problem to perform USB tethering
function.
In this case, usb state of device driver is "DISCONNECTED". So Android does
not recognize correct usb connection state.
Simple way to fix is to remove to disable menu.

Bug: 11176596
Change-Id: Ie6163cf554825e9a1d1c23293338b0e18f5b313d