[Settings] Refresh options menu when showing cached/running processes
While viewing Running processes, if you click on the memory bar in the
free space to show the cached processes, the options menu should change
in line with the different tab content. This patch calls the
onPrepareOptionsMenu in the parent Fragment to correct this.

Issue: https://code.google.com/p/android/issues/detail?id=62350

Change-Id: I89513e8683040edcc2e1d7c4e605e62980c22fb0
Signed-off-by: howardb <android@howardb.com>