Update patch set 3
Patch Set 3: Code-Review-2

Hi, thank-you for your contribution.

Unfortunately this is a change that is actually not desired, and I think would be a serious issue if introduced.

Keep in mind that Android applications are signed by self-signed certificates, not validated by a CA.  So a developer can essentially put whatever they want in the text for their certificate -- the only thing that really matters is the public key, and whether that matches the public key of other apks from the developer.

This means there is nothing we can really show the user about the certificate that they would understand that would be useful, and actually a lot in there could be actively harmful -- the developer could put text in the certificate that makes it sound "safe" like the user might expect from the web experience, but there is in fact no such safety at all.

Patch-set: 3
Reviewer: Dianne Hackborn <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2