Settings: Disable SIM lock UIs in Airplane mode
If the airplane mode is activated using the power button and if the current
activity is one of the SIM lock activities, then the SIM lock UI is not 
disabled. Since the framework doesn't completely support SIM access in airplane
mode, there are few issues seen.

This patch updates sim_lock preference on sim state and airplane mode changes.

Change-Id: Icc9accbadf08d70d92894a58001b89ae6c0dbbba
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: hanpengx <pengfeix.han@intel.com>