Add calibration as a display settings activity
Adds a new "Calibrate" option to the Display
settings which allows users with pen input devices
to correct for parallax and other low-order errors.
The option only appears if a device meeting the
following criteria is present:

 * The device is not virtual
 * The device provides stylus input (SOURCE_STYLUS)
 * The device is direct-input (SOURCE_TOUCHSCREEN)
 * The device has the necessary X and Y axes

After clicking on "Calibrate", the user will be
presented with a set of calibration targets which
respond only to pen input. The device may be re-
oriented at will until the first target is tapped
at which point the orientation is locked for the
remainder of the calibration. Once all of the
targets have been clicked, a test screen appears
where the user can verify the calibration is
sufficiently accurate and either re-calibrate or
accept the calibration.

Change-Id: I26b578655cb7e6e5630a65b2121d9cb313b5b588