Update patch set 1
Patch Set 1:

Thanks, Rob. Gerrit (the code review environment) looks at the Change-Id line in the commit message to see if it's the "same" change or not. I'm not sure what process you followed for modifying the patch, but it looks like your new patch set had/has a new Change-Id and was thus uploaded as a separate change.

In any case, let's move to your new change https://android-review.googlesource.com/#/c/83664/1. I +1'd it. I'm hoping Matthew will review and +2, after which it will be committed/merged.

Patch-set: 1