Settings Res XML: Change battery dishcarging string
The status is shown as "Not charging" in the settings screen when
the charger in unplugged. Modified the strings value so that the
status is shown as "Discharging" once the cable is unplugged.

The change was done to indicate the two different charging status.

Not Charging - Charger Connected, but not charging, charger supplying
power to platform. Battery is not discharging since charger is supplying
power
Discharging - When no charger connected, Battery is discharging.

Change-Id: I07dd97f6691cb869c93406a264db5ebb29de1776
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>