Fix Wifi WPS Pin Entry memory leak (settings app).
In WpsDialog.onStart(), remember the key under which the listener is
registered in the call to WifiManager.startWps(), and then in
WpsDialog.onStop(), pass that listener key to WifiManager.cancelWps()
who will remove it from the cache of listeners. This avoids leaking
the canceled listeners and the WpsDialog objects to which they refer
and objects to which those refer, etc.

Change-Id: I0f50c51f3d2f8f1dfde915af3d27e98fa4ae9cd0
Signed-off-by: Tom Keel <thomas.keel@intel.com>
Signed-off-by: Frederic Danis <frederic.danis@intel.com>