Fix Wifi WPS Pin Entry memory leak (settings app).
In WpsDialog.onStart(), remember the key under which the listener is
registered in the call to WifiManager.startWps(), and then in
WpsDialog.onStop(), pass that listener key to WifiManager.cancelWps()
who will remove it from the cache of listeners. This avoids leaking
the canceled listeners and the WpsDialog objects to which they refer
and objects to which those refer, etc.

This depends upon change I275064bf6061c939bc280a865b0ded37cdc6d07a
and vice-versa.

Change-Id: Ia06776375b44384f5d6261ca7f5b3b16831c3be9
Signed-off-by: Tom Keel <thomas.keel@intel.com>
Signed-off-by: Frederic Danis <frederic.danis@intel.com>