Show true % of total battery usage, not using discharged amount
In Settings / Power management / Battery usage, a list is shown
of the items that have used battery. The percentage is
calculated in relation to discharged amount. If no charging
has been done since last FULL charge, this discharged amount
is basically total capacity minus remaining capacity. But if
(repeated) charging to less than full has occurred, the
discharged amount can be larger than 100, causing item
percentages to also be larger than 100. While the number
shown is correct, it is not intuitive to interpret a percentage
number like that. Instead, we calculate the true percentage
compared to the total sum of battery users. That way, the
items in the list should have a sum of (less than) 100%. It
will be less, because small percentages are not displayed.

Change-Id: I630aded3e566f8feec18424961b3f5ba3e461bc2