Days pop-up shows incorrect language
The language of the days in the days pop-up in
Interruptions is not changed when language is changed.
The wrong language is also shown in the days setting.
The fix removes the static property of the
SimpleDateFormat variables so they are recreated with
the updated locale every time the fragment is recreated.

Change-Id: I861d0786d1791a55b8c8d59246a7b62c75ff64ca