Make CryptKeeper pattern field usable on a phone
Currently the CryptKeepers pattern field is about 10x10mm
on a regular phone device, making it exceedingly difficult
to enter a pattern correctly. With this change the pattern
field fills the display as expected.

This change also removes the IME switcher button as that
is not used for pattern entry.

Change-Id: Iea813918e9e8a1af3ab2bee98ce5fca3cbc5fa39