Update patch set 1
Patch Set 1:

> How does this look on a Nexus 5, 6 or 9?

Dont have a nexus handy unfortunately, will have to come back to you on that. We have a patch for this problem for all current Sony phones though. The change just follows the layout directive in crypt keepers password/pin input fields. The pattern input view itself dont request any particual size and with the current layout I dont see how it can be anything but tiny. There is a special case in the code for tablets, and I suspect it is there because of this issue.

Patch-set: 1
Reviewer: Johan Redestig <1002951@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0