Fix wake lock leak when the BT DockService is not stopped in the correct order.
We were noticing a problem where our device wouldn't go fully into
suspend mode. We found that the "StartingDockService" wake lock was
not being released in all cases. It turned out that some times the
DockService onStartCommand was being invoked multiple times concurrently
(which is fine), but stopSelfResult function was being called in the
wrong order. So some times stopSelfResult would return false, and the
wake lock would leak preventing the device from fully suspending.

The specific chain of events that caused the problem in most cases
was a DOCK_EVENT, which triggers MSG_TYPE_UNDOCKED_TEMPORARY, which
holds the wake lock for a second before it sends the
MSG_TYPE_UNDOCKED_PERMANENT message which eventually calls stopSelfResult
and releases the lock.

But during the one second wait, the onStartCommand was invoked other
times due to android.bluetooth.adapter.action.STATE_CHANGED. The
STATE_CHANGED intent is processed quickly and when it is done it invokes
stopSelfResult() with the most recent startId. Then when the
UNDOCKED_PERMANENT message finally gets processed and calls stopSelfResult()
the return code is false, causing the wake lock leak.

Change-Id: Ibb77964f80962609d1e4ab47c472428911138131