Fix wake lock leak when the BT DockService is not stopped in the correct order.
We were noticing a problem where our device wouldn't go fully into
suspend mode. We found that the "StartingDockService" wake lock was
not being released in all cases. It turned out that some times the
DockService onStartCommand was being invoked multiple times concurrently
(which is fine), but stopSelfResult function was being called in the
wrong order. So some times stopSelfResult would return false, and the
wake lock would leak preventing the device from fully suspending.

The specific chain of events that caused the problem in most cases
was a DOCK_EVENT, which triggers MSG_TYPE_UNDOCKED_TEMPORARY, which
holds the wake lock for a second before it sends the
MSG_TYPE_UNDOCKED_PERMANENT message which eventually calls stopSelfResult
and releases the lock.

But during the one second wait, the onStartCommand was invoked other
times due to android.bluetooth.adapter.action.STATE_CHANGED. The
STATE_CHANGED intent is processed quickly and when it is done it invokes
stopSelfResult() with the most recent startId. Then when the
UNDOCKED_PERMANENT message finally gets processed and calls stopSelfResult()
the return code is false, causing the wake lock leak.

In case you prefer a bullet list:

1) onStartCommand called with startId=1, ACTION_DOCK_EVENT
2) Takes wake lock, then waits for the grace period to send the UNDOCKED_PERMANENT
3) onStartCommand called with startId=2, BluetoothAdapter.ACTION_STATE_CHANGED
4) Takes wake lock again due to BT state change
5) The BT state change finishes and then calls stopSelfResult(startId=2)
6) stopSelfResult is successful and the wake lock is released once
7) The grace period expires and UNDOCKED_PERMANENT finally gets sent
8) UNDOCKED_PERMANENT is processed calls stopSelfResult(startId=1)
9) stopSelfResult is not successful, returns false, and the wake lock is leaked

Change-Id: Ibb77964f80962609d1e4ab47c472428911138131